Refactor Budget components
